Extract operator to its own package, move controller tests to their respective packages
Moving tests should improve coverage accuracy and test caching
