+++ internal/pointer/value.go
@@ -1,0 +1,5 @@
+package pointer
+
+func Of[Value any](value Value) *Value {
+	return &value
+}
