+++ internal/pool/pool.go
@@ -11,6 +11,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/johnstarich/zfs-sync-operator/internal/baddeepcopy"
 	"github.com/johnstarich/zfs-sync-operator/internal/name"
+	"github.com/johnstarich/zfs-sync-operator/internal/pointer"
 	"github.com/johnstarich/zfs-sync-operator/internal/wireguard"
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/ssh"
@@ -131,7 +132,7 @@ func (p Pool) WithSession(ctx context.Context, client ctrlclient.Client, do func
 		hostKeyCallback = ssh.HostKeyCallback(func(_ string, _ net.Addr, key ssh.PublicKey) error {
 			// Save host key for validation in next reconcile
 			poolWithHostKey := p
-			poolWithHostKey.Spec.SSH.HostKey = toPointer(key.Marshal())
+			poolWithHostKey.Spec.SSH.HostKey = pointer.Of(key.Marshal())
 			err := client.Update(ctx, &poolWithHostKey)
 			resourceVersion = poolWithHostKey.ResourceVersion
 			return err
